requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
arrow==1.2.3
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
black==23.3.0
blessed==1.20.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.3
colored==1.4.4
colorlog==6.7.0
contourpy==1.0.7
croniter==1.3.15
cycler==0.11.0
dateutils==0.6.12
deepdiff==6.3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.88.0
filelock==3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.5.0
fvcore==0.1.5.post20221221
google-auth==2.18.1
google-auth-oauthlib==1.0.0
graphviz==0.20.1
grpcio==1.55.0
h11==0.14.0
h5pickle==0.4.2
h5py==3.9.0
huggingface-hub==0.14.1
hydra-core==1.3.2
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
inquirer==3.1.3
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
lit==16.0.5
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
mdurl==0.1.2
metacity==0.6.2
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-ml-py==11.525.112
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
nvitop==1.1.2
oauthlib==3.2.2
omegaconf==2.3.0
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.9.0
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
peppercorn==0.6
Pillow==9.5.0
platformdirs==3.5.1
plotly==5.14.1
portalocker==2.7.0
protobuf==4.23.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.6
pydantic==1.10.8
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
regex==2023.5.5
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.0
tensorflow-hub==0.13.0
termcolor==2.3.0
threadpoolctl==3.1.0
tokenizers==0.13.3
tomli==2.0.1
torch
torch-geometric
torch-tb-profiler
torchmetrics
torchvision
torchviz
tqdm==4.65.0
traitlets==5.9.0
transformers==4.29.2
triton==2.0.0
typing_extensions==4.6.0
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.22.0
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
Werkzeug==2.3.4
yacs==0.1.8
yarl==1.9.2
zipp==3.15.0
